packages/ocr-core/pyproject.toml
[project]
name = "ocr-core"
version = "0.1.0"
requires-python = ">=3.13"
description = "Core benchmarking framework — metrics, data loading, orchestration."
dependencies = [
    "jiwer>=4.0.0",
    "loguru>=0.7.3",
    "numpy>=2.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pillow-heif>=1.3.0",
    "psutil>=7.2.2",
    "pymupdf>=1.27.2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = [
    "nvidia-ml-py>=13.595.45",
    "torch>=2.11.0",
]
chinese = ["opencc-python-reimplemented>=0.1.7"]
optimal-matching = ["scipy>=1.11"]
test = ["pytest>=8.0", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-W ignore::DeprecationWarning:pytz.tzinfo"

[tool.coverage.run]
source = ["ocr_core"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
ocr-core = { path = "../../packages/ocr-core" }
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_core"]

[tool.isort]
profile = "black"
known_first_party = ["ocr_core"]
